adds cart

+++ backend/src/resolvers/Mutation.js
@@ -3,6 +3,7 @@ const jwt = require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
 const { transport, makeANiceEmail } = require('../mail');
+const { hasPermission } = require('../utils');
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
@@ -48,9 +49,14 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // 1. find the item
-    const item = await ctx.db.query.item({ where }, `{ id, title}`);
+    const item = await ctx.db.query.item({ where }, `{ id title user{ id }}`);
     // 2. check if they own that item or have the permission
-    //TODO
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission => ['ADMIN', 'ITEMDELETE'].includes(permission));
+
+    if(!ownsItem && hasPermissions) {
+      throw new Error('You don\'t have permission to do that!');
+    }
     // 3. delete it!
     return ctx.db.mutation.deleteItem({ where }, info);
   },
@@ -167,6 +173,91 @@ const Mutations = {
     });
     // 8. return the new user
     return updatedUser;
+  },
+  async updatePermissions(parent, args, ctx, info) {
+    // 1. check if they are logged in
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in');
+    }
+    // 2. query the current user
+    const currentUser = await ctx.db.query.user(
+      {
+        where: {
+          id: ctx.request.userId,
+        },
+      },
+      info
+    );
+    // 3. check if they have permissions to do this
+      hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+    // 4. update the permissions
+    return ctx.db.mutation.updateUser({
+      data: {
+        permissions: {
+          set: args.permissions,
+        }
+      },
+      where: {
+        id: args.userId
+      },
+    }, info);
+  },
+  async addToCart(parent, args, ctx, info) {
+    // 1. Make sure they are signed in
+    const userId = ctx.request.userId;
+    if(!userId) {
+      throw new Error('You must be signed in!');
+    }
+    // 2. Query the users current cart
+    const [existingCartItem] = await ctx.db.query.cartItems({
+      where: {
+        user: { id: userId },
+        item: { id: args.id },
+      }
+    });
+    // 3. Check if that item is already in their cart
+    if(existingCartItem) {
+      console.log('This item is already in their cart');
+      return ctx.db.mutation.updateCartItem({
+        where: { id: existingCartItem.id },
+        data: { quantity: existingCartItem.quantity + 1 }
+      },
+      info
+      );
+    }
+    // increment by 1 if it is
+    // if not, create a fresh cartItem for that user!
+    return ctx.db.mutation.createCartItem({
+      data: {
+        user: {
+          connect: { id: userId },
+        },
+        item: {
+          connect: { id: args.id }
+        },
+      },
+    }, info
+    );
+  },
+  async removeFromCart(parent, args, ctx, info) {
+    // 1. find cart item
+    const cartItem = await ctx.db.query.cartItem({
+      where: {
+        id: args.id,
+      },
+    },
+    `{id, user { id }}`
+    );
+    // 1.5. make sure we found an item
+    if(!cartItem) throw new Error('No CartItem found!');
+    // 2. make sure they own that cart item
+    if(cartItem.user.id !== ctx.request.userId) {
+      throw new Error('You do not own this cart item');
+    }
+    // 3. delete that cart item
+    return ctx.db.mutation.deleteCartItem({
+      where: { id: args.id },
+    }, info);
   },
 };
